general bug fixes

+++ install.sh
@@ -20,36 +20,40 @@ echo
 
 case "${1:-"install"}" in
   install)
-    install -v -m 755 -D -t "${BINDIR}/" zfsync
-		install -v -D -t "${SYSTEMDDIR}/" zfsync-snapshot.service
-		install -v -D -t "${SYSTEMDDIR}/" zfsync-mirror.service
-		install -v -D -t "${SYSTEMDDIR}/" zfsync-prune.service
-		install -v -D -t "${SYSTEMDDIR}/" zfsync-snapshot.timer
-    install -v -D -t "${DOCDIR}/" inittest.sh
-    install -v -D -t "${DOCDIR}/" README.md
-    install -v -D -t "${DOCDIR}/" install.sh
-    install -v -D -t "${DOCDIR}/" UNLICENSE
-    install -v -D -t "${ETCDIR}/" snapshot.conf.sample
-    install -v -D -t "${COMPLETEDIR}/zfsync" zfsync-completion.bash
+    install -v -d "${BINDIR}"
+    install -v -d "${DOCDIR}"
     install -v -d "${MANDIR}"
-    gzip --to-stdout zfsync.1 > "${MANDIR}/zfsync.1.gz"
+    install -v -d "${ETCDIR}"
+    install -v -d "${COMPLETEDIR}"
+    install -v -m 755 zfsync "${BINDIR}/"
+		install -v -m 644 zfsync-snapshot.service "${SYSTEMDDIR}/"
+		install -v -m 644 zfsync-mirror.service "${SYSTEMDDIR}/"
+		install -v -m 644 zfsync-prune.service "${SYSTEMDDIR}/"
+		install -v -m 644 zfsync-snapshot.timer "${SYSTEMDDIR}/"
+    install -v -m 755 inittest.sh "${DOCDIR}/"
+    install -v -m 644 README.md "${DOCDIR}/"
+    install -v -m 755 install.sh "${DOCDIR}/"
+    install -v -m 644 UNLICENSE "${DOCDIR}/"
+    install -v -m 644 snapshot.conf.sample "${ETCDIR}/"
+    install -v -m 644 zfsync-completion.bash "${COMPLETEDIR}/zfsync"
+    gzip  --to-stdout zfsync.1 > "${MANDIR}/zfsync.1.gz"
+		chmod 644 "${MANDIR}/zfsync.1.gz"
     "${BINDIR}/zfsync" configuser
 
     printf '\nfinished installing zfsync\n\nread the man page for configuration and use instructions\n\n'
-    printf '\nrun `systemctl daemon-reload`\n\n'
     ;;
   uninstall)
-		systemctl disable --now zfsync-snapshot.service || :
-		systemctl disable --now zfsync-mirror.service || :
-		systemctl disable --now zfsync-snapshot.timer || :
     rm -v "${BINDIR}/zfsync"
     rm -v "${MANDIR}/zfsync.1.gz"
     rm -v -rf "${DOCDIR}"
     rm -v -rf "${ETCDIR}"
     rm -v "${COMPLETEDIR}/zfsync"
-		rm -v "${SYSTEMDDIR}/zfsync-snapshot.service"
-		rm -v "${SYSTEMDDIR}/zfsync-mirror.service"
-		rm -v "${SYSTEMDDIR}/zfsync-snapshot.timer"
+    if [ -d "${SYSTEMDDIR}" ]; then
+      rm -v "${SYSTEMDDIR}/zfsync-snapshot.service"
+      rm -v "${SYSTEMDDIR}/zfsync-mirror.service"
+      rm -v "${SYSTEMDDIR}/zfsync-snapshot.timer"
+      [ -z "${DESTDIR}" ] && systemctl daemon-reload # only run if installing into current system
+    fi
     printf '\nfinished uninstalling zfsync\n\n'
     ;;
 esac
